Add primary key constraint name comparison
Change usage of SQLAlchemy's get_primary_keys to get_pk_constraint which
returns a dict with 'constrained_columns' containing the primary key
fields and optionally 'name' with the name of the pk constraint.
Update get_primary_keys_info to compare the entire dicts and apply
ignores on the constraint name rather than the key names if a 'name'
field exists. If 'name' does not exist, the primary key columns are
compared instead.